requirements.txt
pysendy
ShopifyAPI
